+++ src/ecs/components/velocity.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../transform.h"
+#include "transform.h"
 #include "../../engine/gamestate.h"
 #include "../../engine/maths.h"
